requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.1
diff-match-patch==20241021
Django==4.2.10
django-browser-reload==1.17.0
django-cors-headers==4.6.0
django-import-export==4.3.4
django-pwa==2.0.1
django-tailwind==3.8.0
django-unfold==0.46.0
djangorestframework==3.15.2
idna==3.10
mysql-connector-python==9.2.0
mysqlclient==2.2.7
pymemcache==4.0.0
PyMySQL==1.1.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.3
tablib==3.8.0
urllib3==2.3.0




# asgiref==3.8.1
# Django==4.2.10
# django-browser-reload==1.12.1
# django-pwa==1.1.0
# django-tailwind==3.8.0
# mysql-connector-python==8.3.0
# mysqlclient==2.2.4
# pillow
# pymemcache==4.0.0
# PyMySQL==1.1.0
# sqlparse==0.5.0
# tailwind==3.1.5b0
